Separate http server choice from stargate library
