+++ .idea/modules/stargate.iml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module type="RUST_MODULE" version="4">
+  <component name="NewModuleRootManager" inherit-compiler-output="true">
+    <exclude-output />
+    <content url="file://$MODULE_DIR$/../../stargate">
+      <sourceFolder url="file://$MODULE_DIR$/../../stargate/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/../../stargate/integrations/stargate-tide/src" isTestSource="false" />
+    </content>
+    <orderEntry type="inheritedJdk" />
+    <orderEntry type="sourceFolder" forTests="false" />
+  </component>
+</module>
