+++ query-planner-wasm/Cargo.toml
@@ -1,7 +1,6 @@
 [package]
 name = "query-planner-wasm" # this ends up being the name in the published npm package.
 version = "0.0.0-DO.NOT.CHANGE" # The version is managed in the package.json, by Lerna.
-private = true
 authors = ["Apollo <opensource@apollographql.com>"]
 homepage = "https://github.com/apollographql/federation"
 description = "Bridge code written in Rust to Javascript/Typescript, to be internally used by Apollo Gateway. This package is not meant to be independently consumed."
