+++ stargate/integrations/stargate-tide/src/acephei.graphql
